requirements.txt
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click
cryptography==37.0.4
Flask==1.0.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata
itsdangerous==1.1.0
Jinja2==2.10
PyJWT==1.4.2
MarkupSafe==1.1.0
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.20.0
requests
six==1.16.0
SQLAlchemy==1.4.40
urllib3==1.26.11
Werkzeug
zipp
psycopg2





###safe border
# Price prediction requirements
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
pandas>=1.1.4
numpy>=1.18.5
scipy>=1.4.1  
python-math>=0.0.1


# Model -------------------------------------
tensorboard>=2.4.1
prophet>=1.1.1

# Plotting ------------------------------------
seaborn>=0.11.0
matplotlib>=3.2.2
plotly>=5.10.0
ColorStr>=1.3

# Export --------------------------------------
scikit-learn>=1.1.2  # CoreML quantization
tensorflow>=2.10.0  # TFLite export
keras>=2.10.0
openpyxl

